requirements.txt
autopep8==1.6.0
azure-core==1.21.1
azure-storage-blob==12.9.0
camelot-py==0.9.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.1
et-xmlfile==1.1.0
idna==3.3
isodate==0.6.1
msrest==0.6.21
numpy==1.21.5
oauthlib==3.1.1
opencv-python==4.5.4.60
openpyxl==3.0.9
pandas==1.3.5
pdfminer.six==20211012
pycodestyle==2.8.0
pycparser==2.21
PyPDF2==1.26.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
toml==0.10.2
urllib3==1.26.7
